Cleanup: Types
