+++ src/types/events.ts
@@ -84,33 +84,6 @@ export enum EthersObserverEvents {
 
 export type ChainID = number;
 
-/*
- * GraphQL ColonyActionType enum
- */
-export enum ColonyActionType {
-  MintTokens = 'MINT_TOKENS',
-  Payment = 'PAYMENT',
-  CreateDomain = 'CREATE_DOMAIN',
-  EditDomain = 'EDIT_DOMAIN',
-  UnlockToken = 'UNLOCK_TOKEN',
-  MoveFunds = 'MOVE_FUNDS',
-  ColonyEdit = 'COLONY_EDIT',
-  VersionUpgrade = 'VERSION_UPGRADE',
-  EmitDomainReputationPenalty = 'EMIT_DOMAIN_REPUTATION_PENALTY',
-  EmitDomainReputationReward = 'EMIT_DOMAIN_REPUTATION_REWARD',
-  SetUserRoles = 'SET_USER_ROLES',
-  MintTokensMotion = 'MINT_TOKENS_MOTION',
-  CreateDomainMotion = 'CREATE_DOMAIN_MOTION',
-  EditDomainMotion = 'EDIT_DOMAIN_MOTION',
-  VersionUpgradeMotion = 'VERSION_UPGRADE_MOTION',
-  UnlockTokenMotion = 'UNLOCK_TOKEN_MOTION',
-  PaymentMotion = 'PAYMENT_MOTION',
-  MoveFundsMotion = 'MOVE_FUNDS_MOTION',
-  EmitDomainReputationPenaltyMotion = 'EMIT_DOMAIN_REPUTATION_PENALTY_MOTION',
-  EmitDomainReputationRewardMotion = 'EMIT_DOMAIN_REPUTATION_REWARD_MOTION',
-  ColonyEditMotion = 'COLONY_EDIT_MOTION',
-}
-
 // The Filter type doesn't seem to be exported from colony-js
 export type Filter = Parameters<typeof getLogs>[1];
 
@@ -119,20 +92,3 @@ export type NetworkClients =
   | TokenClient
   | AnyColonyClient
   | AnyVotingReputationClient;
-
-/*
- * Contract calls
- */
-export enum ColonyOperations {
-  MintTokens = 'mintTokens',
-  AddDomain = 'addDomain',
-  EditDomain = 'editDomain',
-  Upgrade = 'upgrade',
-  UnlockToken = 'unlockToken',
-  MakePaymentFundedFromDomain = 'makePaymentFundedFromDomain',
-  MoveFundsBetweenPots = 'moveFundsBetweenPots',
-  EmitDomainReputationPenalty = 'emitDomainReputationPenalty',
-  EmitDomainReputationReward = 'emitDomainReputationReward',
-  EditColony = 'editColony',
-  SetUserRoles = 'setUserRoles',
-}
